fix: bad naming, move processRequest to service
