+++ src/api/routes/studentRequest.route.js
@@ -3,14 +3,14 @@ const {
 	getRequestOfStudent,
 	resetStudent,
 	cancelResetStudent,
-	processRequest
-} = require('../controllers/requestStudent.controller');
+	processStudentRequest
+} = require('../controllers/studentRequest.controller');
 const { isAuthenticateUser, authorizeRoles } = require('../middlewares/CheckAuth');
 const { role } = require('../../utils/role');
 
 const router = require('express').Router();
-//* New endpoint. To replace /request/reset/:id && /request/remove/:id
-router.patch('/request/process/:id', isAuthenticateUser, authorizeRoles([role.staff, role.manager]), processRequest);
+//* New endpoint. To replace /request/reset/:id and /request/remove/:id
+router.patch('/request/:id', isAuthenticateUser, authorizeRoles([role.staff, role.manager]), processStudentRequest);
 
 router.post('/request', isAuthenticateUser, authorizeRoles([role.student]), sendRequestToManager);
